requirements.txt
alembic==1.16.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distlib==0.3.9
Faker==37.1.0
filelock==3.18.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
greenlet==3.2.2
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pipenv==2025.0.1
platformdirs==4.3.7
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
setuptools==80.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0
Werkzeug==3.1.3
WTForms==3.2.1
email_validator
